pyproject.toml
[project]
name = "ai-helpers"
version = "0.1.0"
description = "Small CLI for running predefined LLM prompt flows (anki, coverletter, ...)"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "click>=8.1",
    "python-dotenv>=1.0",
]

[project.scripts]
ai = "ai_helpers.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai_helpers"]
